Add alias "git nuke"
This alias deletes both local and remote branches. It's useful for
a gitflow-based workflow, where there are branches frequently created,
merged, and removed.

For now, I've set the delete flag on the local branch to `-d` rather
than `-D`. This choice means that the alias won't accidentally delete an
unmerged branch (i.e. I don't notice that the merge failed and
accidentally delete anyway). I'm not sure that this is the right choice,
but it seems like a reasonable default for now -- even if slightly
ironic for a command named "nuke".