Deactivate conda env if using brew
My current workflow uses both Anaconda and Homebrew. This tends to work
ok, but Homebrew doesn't install packages cleanly when a conda environment is
active because it finds files in usr/local/lib that were instead placed
by Anaconda and may not permenantly be present.

This Stack Overflow link does a nice job elaborating on the issue: https://stackoverflow.com/questions/42859781/best-practices-with-anaconda-and-brew

It suggests a useful utility function for a `.zshrc` that automatically
deactivates a conda environment if one is active when using brew and
then reactivates the environment when the install is finished.