Report the brew command as the commit message
The `brew()` function is monkeypatched to dump to the Brewfile and
commit the changes when installing or uninstalling packages.

Currently, the changes are all committed with the commit message "Update
Brewfile".  This is a terrible commit message.

This change causes the commit to instead use the actual command to
`brew()` (i.e., "install fzf"). This tweak should led to far superior
commit messages.